requirements.txt
azure-core==1.32.0
azure-storage-queue==12.12.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
idna==3.10
isodate==0.7.2
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
